pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "FACSPy"
version = "0.0.4"
authors = [
  { name="Tarik Exner", email="Tarik.Exner@med.uni-heidelberg.de" },
]
description = "Automated Flow Cytometry Analysis"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
#    "anndata>=0.7.4",
#    # numpy needs a version due to #1320
#    "numpy>=1.17.0",
#    "matplotlib>=3.4",
#    "pandas>=1.0",
#    "scipy>=1.4",
#    "seaborn",
#    "h5py>=3",
#    "tqdm",
#    "scikit-learn>=0.24",
#    "statsmodels>=0.10.0rc2",
#    "patsy",
#    "networkx>=2.3",
#    "natsort",
#    "joblib",
#    "numba>=0.41.0",
#    "umap-learn>=0.3.10",
#    "packaging",
#    "session-info",
#    # for getting the stable version
    "scanpy",
    "pandas>=2.0.2",
    "statsmodels>=0.10.0rc2",
    "lxml>=4.9",
    "flowutils",
    "flowio",
    "KDEpy",
    "anytree",
    "statannotations",
    "parc",
    "minisom",
    "phenograph",
    "flowsom-clustering",
    "seaborn>0.12",
    "imblearn"
]

#[project.urls]
#"Homepage" = "http://www.google.de/"
#"Bug Tracker" = "https://github.com/"
